rename module as per suggestion
