+++ crates/networking/p2p/sync.rs
@@ -4,7 +4,7 @@ mod state_sync;
 mod storage_fetcher;
 mod storage_healing;
 mod trie_rebuild;
-mod utils;
+mod fetcher_queue;
 
 use bytecode_fetcher::bytecode_fetcher;
 use ethrex_blockchain::{error::ChainError, BatchBlockProcessingFailure, Blockchain};
