+++ crates/networking/p2p/sync/storage_fetcher.rs
@@ -15,7 +15,7 @@ use tracing::{debug, error};
 use crate::{
     peer_handler::PeerHandler,
     sync::{
-        trie_rebuild::REBUILDER_INCOMPLETE_STORAGE_ROOT, utils::run_queue, MAX_CHANNEL_MESSAGES,
+        trie_rebuild::REBUILDER_INCOMPLETE_STORAGE_ROOT, fetcher_queue::run_queue, MAX_CHANNEL_MESSAGES,
         STORAGE_BATCH_SIZE,
     },
 };
